TECH: Refactor react project structure
