+++ src/react/features/navigation/components/NavigationSideways.js
