--- src/react/features/reducers.js
@@ -1,13 +1,0 @@
-import navigation from './navigation/reducers'
-import lenses from './lenses/reducers'
-import home from './home/reducers'
-import room from './room/reducers'
-import operations from './operations/reducers'
-
-export default {
-  navigation,
-  lenses,
-  home,
-  room,
-  operations,
-}
