+++ src/react/features/room/Room.js
@@ -1,15 +1,14 @@
 import React from 'react'
 import { Heading, Pane } from 'evergreen-ui'
-import { useStore } from '../../store'
-import { useGetRoom, useFocusDiory } from './hooks'
+import { useDispatch } from '../../store'
+import { useFocusDiory } from './hooks'
 import { setFocus } from '../navigation/actions'
 import Diory from '../../components/diories/Diory'
 import Image from '../../components/diories/Image'
 import { getBackgroundImage } from '../../lib/utils'
 
 const useRoom = () => {
-  useGetRoom()
-  const dispatch = useStore()[1]
+  const dispatch = useDispatch()
   const { diory, diorys } = useFocusDiory()
   return {
     diory,
@@ -44,22 +43,24 @@ const Room = () => {
           {diory.text}
         </Heading>
       )}
-      {diorys.slice(0, MAX_NUMBER_OF_DIORYS_PER_VIEW).map(({ diory, onClick }) => (
-        <Diory
-          key={diory.id}
-          diory={diory}
-          onClick={onClick}
-          flex="1 0 240px"
-          height={160}
-          margin={24}
-          elevation={2}
-          alignSelf="center"
-          color="white"
-          fontWeight="bold"
-          background="grey"
-          aria-controls={`panel-${diory.id}`}
-        />
-      ))}
+      {diorys
+        .slice(0, MAX_NUMBER_OF_DIORYS_PER_VIEW)
+        .map(({ diory, onClick }) => (
+          <Diory
+            key={diory.id}
+            diory={diory}
+            onClick={onClick}
+            flex="1 0 240px"
+            height={160}
+            margin={24}
+            elevation={2}
+            alignSelf="center"
+            color="white"
+            fontWeight="bold"
+            background="grey"
+            aria-controls={`panel-${diory.id}`}
+          />
+        ))}
     </Pane>
   )
 }
