--- src/react/hooks/index.js
@@ -1,3 +1,0 @@
-export { useDiory } from './useDiory'
-export { useDiorys } from './useDiorys'
-export { useCompare } from './useCompare'
