--- src/react/hooks/useDiorys.js
@@ -1,8 +1,0 @@
-import { useStore } from '../store'
-
-export const useDiorys = (ids = {}) => {
-  const [{ diograph }] = useStore(state => state.room)
-  return Object.entries(ids)
-    .map(([key, { id }]) => ({ key, ...diograph[id] }))
-    .filter(({ id }) => id)
-}
