+++ src/react/store/index.js
@@ -1,8 +1,8 @@
-import reducers from '../features/reducers'
+import reducers from './reducers'
 import combineReducers from './combineReducers'
 
 export const reducer = combineReducers(reducers)
 
-export { default as initialState } from '../features/initialState'
+export { default as initialState } from './initialState'
 
 export { StoreProvider, useStore, useDispatch } from './StoreContext'
