+++ src/react/store/reducers.js
@@ -1,0 +1,15 @@
+import navigation from '../features/navigation/reducers'
+import lenses from '../features/lenses/reducers'
+import home from '../features/home/reducers'
+import room from '../features/room/reducers'
+import operations from '../features/operations/reducers'
+import connector from '../features/connector/reducers'
+
+export default {
+  navigation,
+  operations,
+  lenses,
+  room,
+  home,
+  connector,
+}
