+++ src/react/utils/middleware/addMiddleware.js
