+++ src/react/utils/middleware/logger.js
