+++ src/react/utils/scripts/flattenDiograph.js
