+++ src/react/utils/scripts/geoFromDataToDiory.js
