+++ src/react/utils/scripts/index.js
